+++ include/remy_robot_control/types.h
@@ -14,15 +14,15 @@
 
 namespace remy_robot_control {
 
-typedef Eigen::Matrix<double, 6, 3> JacobM;
-typedef Eigen::Matrix<double, 3, 3> JacobMP;
-typedef Angle<double> Angled;
+typedef Eigen::Matrix<float, 6, 3> JacobM;
+typedef Eigen::Matrix<float, 3, 3> JacobMP;
+typedef Angle<float> Angled;
 
 struct RemyJoints {
   Angled q1;
   Angled q2;
   Angled q3;
-  RemyJoints(double q1mM[2], double q2mM[2], double q3mM[2]) {
+  RemyJoints(float q1mM[2], float q2mM[2], float q3mM[2]) {
     q1 = Angled(0, q1mM[0], q1mM[1]);
     q2 = Angled(0, q2mM[0], q2mM[1]);
     q3 = Angled(0, q3mM[0], q3mM[1]);
